Change the way MonitorThread and DatabaseThread works
Before, everytime an event was occured in Docker, we create a new Database thread and we started it. From now on, we start the database thread on the Main.java and the createAllTables() is executed. Then, every time an event occurs, the method keepTrackOf...() of database thread is executed. If the event is a delete (of an instance or image or volume), we call delete...(). In this way, it is much cleaner and more efficient