Create new custom exception for error handling in IndividualContainerController
Also add javadoc and comments on IndividualContainerController, create a new endpoint on our Watchdog Rest Api in which we get the current state of only one instance, and it is very useful in this class because we need it for the cpuchart. Previously, we were getting all instances and comparing every id with ours so to take the specific id. Obviously, that was underproductive adn inefficient. gm lfg